chatbot-kpbu/requirements.txt
pinecone[grpc]
google-generativeai
langchain
langchain-google-genai
langchain-text-splitters
pypdf
python-docx
python-dotenv
